Implement POC for GCS Fuse Executable-Sourced Credentials with C++ dependencies
