--- poc/option1/Dockerfile.mock-service
@@ -1,4 +1,0 @@
-FROM alpine:latest
-COPY mock-service /usr/bin/mock-service
-RUN chmod +x /usr/bin/mock-service
-ENTRYPOINT ["/usr/bin/mock-service"]
