--- poc/option1/Dockerfile.token-fetcher
@@ -1,3 +1,0 @@
-FROM alpine:latest
-COPY fetch-token /usr/bin/fetch-token
-RUN chmod +x /usr/bin/fetch-token
