+++ poc/option1/base_demo/mock-credential-configuration.json
@@ -1,0 +1,15 @@
+{
+  "type": "external_account",
+  "audience": "//iam.googleapis.com/projects/332244367552/locations/global/workloadIdentityPools/custom-simulation-pool/providers/custom-saml-provider",
+  "subject_token_type": "urn:ietf:params:oauth:token-type:saml2",
+  "token_url": "http://mock-token-service:5000/sts-token",
+  "credential_source": {
+    "executable": {
+      "command": "/scripts/fetch-token",
+      "timeout_millis": 5000,
+      "output_format": {
+        "type": "json"
+      }
+    }
+  }
+}
