+++ poc/option1/cpp_poc/Dockerfile
@@ -1,0 +1,2 @@
+FROM alpine
+COPY mock-client-static /usr/bin/mock-client-static
