+++ poc/option1/cpp_poc/cpp-credential-configuration.json
@@ -1,0 +1,13 @@
+{
+  "type": "external_account",
+  "audience": "//iam.googleapis.com/projects/332244367552/locations/global/workloadIdentityPools/meta-simulation-pool/providers/meta-saml-provider",
+  "subject_token_type": "urn:ietf:params:oauth:token-type:saml2",
+  "token_url": "https://sts.googleapis.com/v1/token",
+  "credential_source": {
+    "executable": {
+      "command": "/scripts/mock-client-static",
+      "timeout_millis": 30000,
+      "output_file": "/tmp/token-output.json"
+    }
+  }
+}
