+++ poc/option1/cpp_poc/main.rs
@@ -1,0 +1,11 @@
+extern "C" {
+    fn mock_encrypt();
+}
+
+fn main() {
+    println!("Rust wrapper starting...");
+    unsafe {
+        mock_encrypt();
+    }
+    println!("Rust wrapper finished.");
+}
