+++ poc/option1/cpp_poc/mock_crypto.cpp
@@ -1,0 +1,7 @@
+#include <iostream>
+
+extern "C" {
+    void mock_encrypt() {
+        std::cout << "MOCK CRYPTO: Encrypting using C++ standard library (cout)..." << std::endl;
+    }
+}
