--- poc/option1/hostpath-deployment.yaml
@@ -1,36 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: gke-workload-with-gcsfuse-hostpath
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: gke-workload-with-gcsfuse-hostpath
-  template:
-    metadata:
-      labels:
-        app: gke-workload-with-gcsfuse-hostpath
-      annotations:
-        gke-gcsfuse/volumes: "true"
-        gke-gcsfuse/workload-identity-credential-configmap: "mock-auth-config"
-        gke-gcsfuse/additional-volume-mounts: "binary-volume:/scripts"
-    spec:
-      containers:
-      - name: workload
-        image: ubuntu
-        command: ["sleep", "infinity"]
-        volumeMounts:
-        - name: demo-bucket-vol
-          mountPath: /data
-      
-      volumes:
-      - name: binary-volume
-        hostPath:
-          path: /var/lib/custom/bin # Example path where the binary is dropped
-          type: DirectoryOrCreate
-      - name: demo-bucket-vol
-        csi:
-          driver: gcsfuse.csi.storage.gke.io
-          volumeAttributes:
-            bucketName: yangspirit-joonix-demo-bucket
